Added install action and notion of application/database versions to allow application writers to include update code.
Now, a file conf/Installer.php with a class conf_Installer can be
included with the application.  This class can have methods of the 
form update_xxx() where xxx is the version number.  If the file system
is updated to a new version, a message will be displayed in the application
alerting the user that the database is out of date.

It includes an install action to perform the updates. This action runs
all methods update_xxx() where xxx > database version and xxx <= file system version.

The database version is stored in a single row table named database__version.

The file system version is stored in a text file named version.txt however the
actual version number is interpreted as the last integer in this file.

e.g. if version.txt contains 0.5 beta 2 2345
then the version is 2345
This allows us to have a running version number that is an integer which makes
updates easier.