Added setAuthType() method to Authentication tool to be able to change the authentication type mid request.
Also added auth_type directive to the actions.ini file to specify a certain type of authentication to be used for a particular action.

This should be used with care because if an action is included in the application specifying an authentication type that isn't supported, then problems may arise - i.e. security holes.

Best practice is for the application administrator to set this for his own actions, rather than including this directive in dataface actions.  However certain modules may find this advantageous to use.

This was really created to allow us to specify http authentication for things like export_xml, export_json, fmpxmlexport and other actions that will most likely be run by machine.