Added a Messages module to be able to send messages around a Dataface application.  Messages are useful to allow users to negotiate the terms of data entry before anything is persisted to the database.
A simple example of this is an application with users and groups where a user may apply to be added to a group but he must first be approved.  In this case we can't simply have the user add himself via the add related record option.  Instead he sends a "Message" to the group owner with his join request.  The group owner would then approve or reject the user.

This module is meant to contain all of the guts necessary to abstract the message passing process so that a minimal amount of coding is required on a per-application basis.