Fixed bug 128 http://bugs.weblite.ca/view.php?id=128
Which deals with group widgets on related record forms.
In fact there were greater problems with group widgets that
caused them to be rendered twice on edit forms.  This has also been fixed.

This update also adds a template parameter for field groups to allow the developer to provide a custom template for a section of the edit form.