Changed behavior of permissions so that field__permissions() and rel_xxx__permissions() are treated as partial permission definitions. These are merged against teh record permissions to obtain the full resolution.
This allows us to make a call to 
$record->checkPermission('view related records', array('relationship'=>'abstracts');
to truly check if the user is allowed to view the abstracts relationship.

Then the rel_abstracts__permissions() method need only add the changes to the defaults that need to be made.

e.g.
function rel_abstracts__permissions(&$record){
    return array("view related records"=>0);
}

THis would disallow teh view related records permission for that record even if all of the other relationships in the record van be viewed.

So where is this behavior different than the previous?


Well in the above example we simply defined that the view related records permission was 0 for that relationship.  The old behavior would have set all permissions 0 because of their absence in the return value of rel_abstracts__permissions().  Now the result of this method is simply merged against the default permissions (from getPermissions()), so that we only need to specify differences.