Fixed critical issue with caching and InnoDB tables. Since InnoDB tables don't store update time (http://bugs.mysql.com/bug.php?id=14374) getTableModificationTimes() was returning 0 for update time which would mean that the cache would think that the table hasn't been updated. Changed so that the default is the current time - however this will reduce the effectiveness of the cache if you have a lot of InnoDB tables.
Conclusion:  For performance MyISAM tables are much better.  Though for data integrity it may be necessary to use InnoDB on some larger database projects.

If you have a front end that is read only, perhaps you may want to have a duplicate database with MyISAM tables for that one that is synchronized periodically.