Added feature 583.  The ability to apply decorators to valuelists in the delegate class.
This allows us to modify the values or labels of a valuelist in the delegate class by defining a method:

valuelist__[valuelistname]__decorate($values)

which takes the existing valuelist values as a parameter.  This list can be modified then returned.

This is most useful for internationalization of valuelists so that we can modify certain labels.

e.g.

function valuelist__languages__decorate($languages){
    $app =& Dataface_Application::getInstance();
    if ( $app->_conf['lang'] == 'fr' ){
        $languages['en'] = 'Anglais';
    }
    return $languages;
}

Or something along those lines.  It could also return an entirely new array.